Refactoring per review comments.
